Fix Number.is* to accept unknown again
This essentially reapplies #24436 which seems to have been accidentally reverted by 2f73986b44a4ec10c7ebe7188e23863c3879d54e.

Note that this is specifically _not_ the same issue as #4002, which pertains to the global symbols that coerce their arguments, and should therefore _not_ accept `unknown`.  These functions always return `false` for any argument that is not strictly a `number`, so it is significantly safer to accept all arguments.